change for double

+++ include/frac/shader.hpp
@@ -12,7 +12,7 @@ public:
 	Shader(const std::string& frag_shader_path, const std::string& vert_shader_path);
 	GLuint get_compiled_shader();
 	void set_float(const std::string& uniform_name, float val);
-	void set_vec2(const std::string& uniform_name, const std::vector<float>& val);
+	void set_vec2(const std::string& uniform_name, const std::vector<double>& val);
 private:
 	void compile_shader();
